feat: Implementations of Monitor for system resources and process activity
Atop is an ASCII full-screen performance monitor, similar to the top command, but atop only shows the active system-resources and processes, and only shows the deviations since the previous interval. At regular intervals, it shows system-level activity related to the CPU, memory, swap, disks and network layers, and it shows for every active process the CPU utilization in system and user mode, the virtual and resident memory growth, priority, username, state, and exit code. The process level activity is also shown for processes which finished during the last interval, to get a complete overview about the consumers of things such as CPU time.

Log: Implementations of Monitor for system resources and process activity

Issue: https://github.com/deepin-community/sig-deepin-sysdev-team/issues/241